refactored ace.system.remote ns
